+++ guest/rust/examples/games/pong/src/server.rs
@@ -4,7 +4,7 @@ use ambient_api::{
     core::{
         messages::Frame,
         physics::components::linear_velocity,
-        player::components::{player, user_id},
+        player::components::{is_player, user_id},
         primitives::{components::cube, concepts::make_sphere},
         rendering::components::color,
         transform::{components::*, concepts::make_transformable},
@@ -64,7 +64,7 @@ pub fn main() {
         .spawn();
 
     // When a player spawns, create a camera and other components for them
-    spawn_query(player()).bind(move |players| {
+    spawn_query(is_player()).bind(move |players| {
         for (player, _) in players {
             entity::add_component(player, player_movement_direction(), 0.0);
         }
@@ -72,7 +72,7 @@ pub fn main() {
 
     // When a player despawns, clean up their objects
     let player_objects_query = query(user_id()).build();
-    despawn_query(user_id()).requires(player()).bind({
+    despawn_query(user_id()).requires(is_player()).bind({
         move |players| {
             let player_objects = player_objects_query.evaluate();
             for (_, player_user_id) in &players {
@@ -107,7 +107,7 @@ pub fn main() {
     });
 
     Frame::subscribe(move |_| {
-        let players = entity::get_all(player());
+        let players = entity::get_all(is_player());
 
         // start the ball if we have 2 players and ball has no velocity
         if players.len() >= 2 && entity::get_component(ball, linear_velocity()).is_none() {